+++ tests/test_model_v2.py
@@ -1,12 +1,12 @@
 """Tests for OData Model module"""
 # pylint: disable=line-too-long,too-many-locals,too-many-statements,invalid-name, too-many-lines, no-name-in-module, expression-not-assigned, pointless-statement
 import os
-from datetime import datetime, timezone
+from datetime import datetime, timezone, timedelta
 from unittest.mock import patch
 import pytest
 from pyodata.v2.model import Schema, Typ, StructTypeProperty, Types, EntityType, EdmStructTypeSerializer, \
     Association, AssociationSet, EndRole, AssociationSetEndRole, TypeInfo, MetadataBuilder, ParserError, PolicyWarning, \
-    PolicyIgnore, Config, PolicyFatal, NullType, NullAssociation, current_timezone, StructType
+    PolicyIgnore, Config, PolicyFatal, NullType, NullAssociation, current_timezone, StructType, parse_datetime_literal
 from pyodata.exceptions import PyODataException, PyODataModelError, PyODataParserError
 from tests.conftest import assert_logging_policy
 import pyodata.v2.model
@@ -472,6 +472,32 @@ def test_traits():
     assert str(e_info.value).startswith("Malformed value '1234-56' for primitive")
 
 
+@pytest.mark.parametrize('datetime_literal,expected', [
+    ('2001-02-03T04:05:06.000007', datetime(2001, 2, 3, 4, 5, 6, microsecond=7)),
+    ('2001-02-03T04:05:06', datetime(2001, 2, 3, 4, 5, 6, 0)),
+    ('2001-02-03T04:05', datetime(2001, 2, 3, 4, 5, 0, 0)),
+])
+def test_parse_datetime_literal(datetime_literal, expected):
+    assert parse_datetime_literal(datetime_literal) == expected
+
+
+@pytest.mark.parametrize('illegal_input', [
+    '2001-02-03T04:05:61',
+    '2001-02-03T04:61',
+    '2001-02-03T24:05',
+    '2001-02-32T04:05',
+    '2001-13-03T04:05',
+    '2001-00-03T04:05',
+    '01-02-03T04:05',
+    '2001-02-03T04:05.AAA',
+    '',
+])
+def test_parse_datetime_literal_faulty(illegal_input):
+    with pytest.raises(PyODataModelError) as e_info:
+        parse_datetime_literal(f'{illegal_input}')
+    assert str(e_info.value).startswith(f'Cannot decode datetime from value {illegal_input}')
+
+
 def test_traits_datetime():
     """Test Edm.DateTime traits"""
 
@@ -613,6 +639,153 @@ def test_traits_datetime():
     with pytest.raises(PyODataModelError) as e_info:
         typ.traits.from_json("/Date(xyz)/")
     assert str(e_info.value).startswith('Cannot decode datetime from value xyz')
+
+
+def test_traits_datetimeoffset(type_date_time_offset):
+    """Test Edm.DateTimeOffset traits"""
+
+    assert repr(type_date_time_offset.traits) == 'EdmDateTimeOffsetTypTraits'
+
+
+def test_traits_datetimeoffset_to_literal(type_date_time_offset):
+    """Test Edm.DateTimeOffset trait: Python -> literal"""
+
+    testdate = datetime(2005, 1, 28, 18, 30, 44, 123456, tzinfo=timezone(timedelta(hours=3, minutes=40)))
+    assert type_date_time_offset.traits.to_literal(testdate) == "datetimeoffset'2005-01-28T18:30:44.123456+03:40'"
+
+    # without milliseconds part, negative offset
+    testdate = datetime(2005, 1, 28, 18, 30, 44, 0, tzinfo=timezone(-timedelta(minutes=100)))
+    assert type_date_time_offset.traits.to_literal(testdate) == "datetimeoffset'2005-01-28T18:30:44-01:40'"
+
+
+def test_traits_invalid_datetimeoffset_to_literal(type_date_time_offset):
+    # serialization of invalid value
+    with pytest.raises(PyODataModelError) as e_info:
+        type_date_time_offset.traits.to_literal('xyz')
+    assert str(e_info.value).startswith('Cannot convert value of type')
+
+
+@pytest.mark.parametrize('python_datetime,expected,comment', [
+    (datetime(1976, 11, 23, 3, 33, 6, microsecond=123000, tzinfo=timezone.utc), '/Date(217567986123+0000)/', 'UTC'),
+    (datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=14))), '/Date(217567986000+0840)/', '+14 hours'),
+    (datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=-12))), '/Date(217567986000-0720)/', '-12 hours'),
+    ])
+def test_traits_datetimeoffset_to_json(type_date_time_offset, python_datetime, expected, comment):
+    """Test Edm.DateTimeOffset trait: Python -> json"""
+
+    assert type_date_time_offset.traits.to_json(python_datetime) == expected, comment
+
+
+@pytest.mark.parametrize('literal,expected,comment', [
+    ("datetimeoffset'1976-11-23T03:33:06.654321+12:11'",
+     datetime(1976, 11, 23, 3, 33, 6, microsecond=654321, tzinfo=timezone(timedelta(hours=12, minutes=11))),
+     'Full representation'),
+    ("datetimeoffset'1976-11-23T03:33:06+12:11'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=12, minutes=11))), 'No milliseconds'),
+    ("datetimeoffset'1976-11-23T03:33:06-01:00'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=-1))), 'Negative offset'),
+    ("datetimeoffset'1976-11-23t03:33:06-01:00'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=-1))), "lowercase 'T' is valid"),
+    ("datetimeoffset'1976-11-23T03:33:06+00:00'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone.utc), '+00:00 is UTC'),
+    ("datetimeoffset'1976-11-23T03:33:06-00:00'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone.utc), '-00:00 is UTC'),
+    ("datetimeoffset'1976-11-23t03:33:06Z'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone.utc), 'Z is UTC'),
+    ("datetimeoffset'1976-11-23t03:33:06+12:00'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=12))), 'On dateline'),
+    ("datetimeoffset'1976-11-23t03:33:06-12:00'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=-12))), 'Minimum offset'),
+    ("datetimeoffset'1976-11-23t03:33:06+14:00'", datetime(1976, 11, 23, 3, 33, 6, tzinfo=timezone(timedelta(hours=14))), 'Maximum offset'),
+])
+def test_traits_datetimeoffset_from_literal(type_date_time_offset, literal, expected, comment):
+    """Test Edm.DateTimeOffset trait: literal -> Python"""
+
+    assert expected == type_date_time_offset.traits.from_literal(literal), comment
+
+
+def test_traits_datetimeoffset_from_invalid_literal(type_date_time_offset):
+    with pytest.raises(PyODataModelError) as e_info:
+        type_date_time_offset.traits.from_literal('xyz')
+    assert str(e_info.value).startswith('Malformed value xyz for primitive')
+
+    with pytest.raises(PyODataModelError) as e_info:
+        type_date_time_offset.traits.from_literal("datetimeoffset'xyz'")
+    assert str(e_info.value).startswith('Cannot decode datetimeoffset from value xyz')
+
+
+def test_traits_datetimeoffset_from_odata(type_date_time_offset):
+    """Test Edm.DateTimeOffset trait: OData -> Python"""
+
+    # parsing full representation
+    testdate = type_date_time_offset.traits.from_json("/Date(217567986010+0060)/")
+    assert testdate.year == 1976
+    assert testdate.month == 11
+    assert testdate.day == 23
+    assert testdate.hour == 3
+    assert testdate.minute == 33
+    assert testdate.second == 6
+    assert testdate.microsecond == 10000
+    assert testdate.tzinfo == timezone(timedelta(hours=1))
+
+    # parsing without milliseconds, negative offset
+    testdate = type_date_time_offset.traits.from_json("/Date(217567986000-0005)/")
+    assert testdate.year == 1976
+    assert testdate.minute == 33
+    assert testdate.second == 6
+    assert testdate.microsecond == 0
+    assert testdate.tzinfo == timezone(-timedelta(minutes=5))
+
+    # parsing below lowest value with workaround
+    pyodata.v2.model.FIX_SCREWED_UP_MINIMAL_DATETIME_VALUE = True
+    testdate = type_date_time_offset.traits.from_json("/Date(-62135596800001+0001)/")
+    assert testdate.year == 1
+    assert testdate.month == 1
+    assert testdate.day == 1
+    assert testdate.minute == 0
+    assert testdate.tzinfo == timezone(timedelta(minutes=1))
+
+    # parsing the lowest value
+    pyodata.v2.model.FIX_SCREWED_UP_MINIMAL_DATETIME_VALUE = False
+    with pytest.raises(PyODataModelError) as e_info:
+        type_date_time_offset.traits.from_json("/Date(-62135596800001+0001)/")
+    assert str(e_info.value).startswith('Cannot decode datetime from value -62135596800001.')
+
+    testdate = type_date_time_offset.traits.from_json("/Date(-62135596800000+0055)/")
+    assert testdate.year == 1
+    assert testdate.month == 1
+    assert testdate.day == 1
+    assert testdate.hour == 0
+    assert testdate.minute == 0
+    assert testdate.second == 0
+    assert testdate.microsecond == 0
+    assert testdate.tzinfo == timezone(timedelta(minutes=55))
+
+    # parsing above highest value with workaround
+    pyodata.v2.model.FIX_SCREWED_UP_MAXIMUM_DATETIME_VALUE = True
+    testdate = type_date_time_offset.traits.from_json("/Date(253402300800000+0055)/")
+    assert testdate.year == 9999
+    assert testdate.month == 12
+    assert testdate.day == 31
+    assert testdate.minute == 0
+    assert testdate.tzinfo == timezone(timedelta(minutes=55))
+
+    # parsing the highest value
+    pyodata.v2.model.FIX_SCREWED_UP_MAXIMUM_DATETIME_VALUE = False
+    with pytest.raises(PyODataModelError) as e_info:
+        type_date_time_offset.traits.from_json("/Date(253402300800000+0055)/")
+    assert str(e_info.value).startswith('Cannot decode datetime from value 253402300800000.')
+
+    testdate = type_date_time_offset.traits.from_json("/Date(253402300799999-0001)/")
+    assert testdate.year == 9999
+    assert testdate.month == 12
+    assert testdate.day == 31
+    assert testdate.hour == 23
+    assert testdate.minute == 59
+    assert testdate.second == 59
+    assert testdate.microsecond == 999000
+    assert testdate.tzinfo == timezone(-timedelta(minutes=1))
+
+    # parsing invalid value
+    with pytest.raises(PyODataModelError) as e_info:
+        type_date_time_offset.traits.from_json("xyz")
+    assert str(e_info.value).startswith('Malformed value xyz for primitive')
+
+    with pytest.raises(PyODataModelError) as e_info:
+        type_date_time_offset.traits.from_json("/Date(xyz)/")
+    assert str(e_info.value).startswith('Malformed value /Date(xyz)/ for primitive Edm.DateTimeOffset type.')
 
 
 def test_traits_collections():